requirements.txt
streamlit
requests
beautifulsoup4
pandas
selenium
webdriver-manager
googletrans==4.0.0-rc1
pillow
